Do not repeat featured blog post in list
On the home page and on the blog list page, we show the
latest article as big and featured and show all other
articles below.
In the current version, the latest article *also* gets
shown in the list below which is redundant. This change
hence removes it from the list below.