Tweak `mixed` tile style
Project tiles can be marked up as having a `mixed` background. For those, we used to add as background shadow for white text. With this change, we add a shadow in all direction which looks like an outline. We also change the text to black and the outline to white. This should increase readability for those tiles.